@@ -1,26 +1,19 @@
-cmake_minimum_required(VERSION 2.8.3)
+cmake_minimum_required(VERSION 3.10) 
 project(linefit_ground_segmentation_ros)
 
-find_package(catkin_simple 0.1.0 REQUIRED )
+# find_package(catkin_simple 0.1.0 REQUIRED )
 
-catkin_simple(ALL_DEPS_REQUIRED)
+# catkin_simple(ALL_DEPS_REQUIRED)
+set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
+add_definitions(-std=c++17 -O3)
 
-add_definitions(-std=c++14)
+find_package(ament_cmake_auto)
 
-find_package(PCL REQUIRED)
-include_directories(${PCL_INCLUDE_DIRS})
-link_directories(${PCL_LIBRARY_DIRS})
-add_definitions(${PCL_DEFINITIONS})
+ament_auto_find_build_dependencies()
+ament_auto_add_executable(ground_segmentation_node src/ground_segmentation_node.cc)
+ament_auto_add_executable(ground_segmentation_test_node src/ground_segmentation_test_node.cc)
 
-add_executable(ground_segmentation_node src/ground_segmentation_node.cc)
-add_executable(ground_segmentation_test_node src/ground_segmentation_test_node.cc)
-target_link_libraries(ground_segmentation_node ${catkin_LIBRARIES} ${PCL_LIBRARIES})
-target_link_libraries(ground_segmentation_test_node ${catkin_LIBRARIES} ${PCL_LIBRARIES})
-
-#add_doxygen(NOT_AUTOMATIC)
-
-cs_install()
-cs_export()
+ament_auto_package(INSTALL_TO_SHARE launch)
 
 #############
 # QTCREATOR #
