@@ -1,115 +1,155 @@
-#include <eigen_conversions/eigen_msg.h>
-#include <ros/ros.h>
+#include <memory>
+
+#include <geometry_msgs/msg/transform_stamped.hpp>
 #include <pcl/common/transforms.h>
 #include <pcl/io/ply_io.h>
 #include <pcl_conversions/pcl_conversions.h>
-#include <pcl_ros/point_cloud.h>
+// #include <pcl_ros/point_cloud.hpp>
+#include <rclcpp/qos.hpp>
+#include <rclcpp/rclcpp.hpp>
+#include <sensor_msgs/msg/point_cloud2.hpp>
+#include <tf2_ros/buffer.h>
 #include <tf2_ros/transform_listener.h>
 
 #include "ground_segmentation/ground_segmentation.h"
 
-class SegmentationNode {
-  ros::Publisher ground_pub_;
-  ros::Publisher obstacle_pub_;
-  tf2_ros::Buffer tf_buffer_;
-  tf2_ros::TransformListener tf_listener_{tf_buffer_};
-  GroundSegmentationParams params_;
-  GroundSegmentation segmenter_;
-  std::string gravity_aligned_frame_;
-
+class SegmentationNode : public rclcpp::Node {
 public:
-  SegmentationNode(ros::NodeHandle& nh,
-                   const std::string& ground_topic,
-                   const std::string& obstacle_topic,
-                   const GroundSegmentationParams& params,
-                   const bool& latch = false) : params_(params), segmenter_(params) {
-    ground_pub_ = nh.advertise<pcl::PointCloud<pcl::PointXYZ>>(ground_topic, 1, latch);
-    obstacle_pub_ = nh.advertise<pcl::PointCloud<pcl::PointXYZ>>(obstacle_topic, 1, latch);
-    nh.param<std::string>("gravity_aligned_frame", gravity_aligned_frame_, "");
-  }
+  SegmentationNode(const rclcpp::NodeOptions &node_options);
+  void scanCallback(const sensor_msgs::msg::PointCloud2::SharedPtr msg);
 
-  void scanCallback(const pcl::PointCloud<pcl::PointXYZ>& cloud) {
-    pcl::PointCloud<pcl::PointXYZ> cloud_transformed;
-
-    std::vector<int> labels;
-
-    bool is_original_pc = true;
-    if (!gravity_aligned_frame_.empty()) {
-      geometry_msgs::TransformStamped tf_stamped;
-      try{
-        tf_stamped = tf_buffer_.lookupTransform(gravity_aligned_frame_, cloud.header.frame_id,
-                                                pcl_conversions::fromPCL(cloud.header.stamp));
-        // Remove translation part.
-        tf_stamped.transform.translation.x = 0;
-        tf_stamped.transform.translation.y = 0;
-        tf_stamped.transform.translation.z = 0;
-        Eigen::Affine3d tf;
-        tf::transformMsgToEigen(tf_stamped.transform, tf);
-        pcl::transformPointCloud(cloud, cloud_transformed, tf);
-        is_original_pc = false;
-      }
-      catch (tf2::TransformException &ex) {
-        ROS_WARN_THROTTLE(1.0, "Failed to transform point cloud into gravity frame: %s",ex.what());
-      }
-    }
-
-    // Trick to avoid PC copy if we do not transform.
-    const pcl::PointCloud<pcl::PointXYZ>& cloud_proc = is_original_pc ? cloud : cloud_transformed;
-
-    segmenter_.segment(cloud_proc, &labels);
-    pcl::PointCloud<pcl::PointXYZ> ground_cloud, obstacle_cloud;
-    ground_cloud.header = cloud.header;
-    obstacle_cloud.header = cloud.header;
-    for (size_t i = 0; i < cloud.size(); ++i) {
-      if (labels[i] == 1) ground_cloud.push_back(cloud[i]);
-      else obstacle_cloud.push_back(cloud[i]);
-    }
-    ground_pub_.publish(ground_cloud);
-    obstacle_pub_.publish(obstacle_cloud);
-  }
+  rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr ground_pub_;
+  rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr obstacle_pub_;
+  std::shared_ptr<tf2_ros::Buffer> tf_buffer_;
+  std::shared_ptr<tf2_ros::TransformListener> tf_listener_;
+  rclcpp::Subscription<sensor_msgs::msg::PointCloud2>::SharedPtr cloud_sub_;
+  GroundSegmentationParams params_;
+  std::shared_ptr<GroundSegmentation> segmenter_;
+  std::string gravity_aligned_frame_;
 };
 
-int main(int argc, char** argv) {
-  ros::init(argc, argv, "ground_segmentation");
+SegmentationNode::SegmentationNode(const rclcpp::NodeOptions &node_options)
+    : Node("ground_segmentation", node_options) {
+  gravity_aligned_frame_ =
+      this->declare_parameter("gravity_aligned_frame", "gravity_aligned");
 
-  ros::NodeHandle nh("~");
-
-  // Do parameter stuff.
-  GroundSegmentationParams params;
-  nh.param("visualize", params.visualize, params.visualize);
-  nh.param("n_bins", params.n_bins, params.n_bins);
-  nh.param("n_segments", params.n_segments, params.n_segments);
-  nh.param("max_dist_to_line", params.max_dist_to_line, params.max_dist_to_line);
-  nh.param("max_slope", params.max_slope, params.max_slope);
-  nh.param("min_slope", params.min_slope, params.min_slope);
-  nh.param("long_threshold", params.long_threshold, params.long_threshold);
-  nh.param("max_long_height", params.max_long_height, params.max_long_height);
-  nh.param("max_start_height", params.max_start_height, params.max_start_height);
-  nh.param("sensor_height", params.sensor_height, params.sensor_height);
-  nh.param("line_search_angle", params.line_search_angle, params.line_search_angle);
-  nh.param("n_threads", params.n_threads, params.n_threads);
+  params_.visualize = this->declare_parameter("visualize", params_.visualize);
+  params_.n_bins = this->declare_parameter("n_bins", params_.n_bins);
+  params_.n_segments =
+      this->declare_parameter("n_segments", params_.n_segments);
+  params_.max_dist_to_line =
+      this->declare_parameter("max_dist_to_line", params_.max_dist_to_line);
+  params_.max_slope = this->declare_parameter("max_slope", params_.max_slope);
+  params_.min_slope = this->declare_parameter("min_slope", params_.min_slope);
+  params_.long_threshold =
+      this->declare_parameter("long_threshold", params_.long_threshold);
+  params_.max_long_height =
+      this->declare_parameter("max_long_height", params_.max_long_height);
+  params_.max_start_height =
+      this->declare_parameter("max_start_height", params_.max_start_height);
+  params_.sensor_height =
+      this->declare_parameter("sensor_height", params_.sensor_height);
+  params_.line_search_angle =
+      this->declare_parameter("line_search_angle", params_.line_search_angle);
+  params_.n_threads = this->declare_parameter("n_threads", params_.n_threads);
+  params_.r_min = this->declare_parameter("r_min", params_.r_min);
+  params_.r_max = this->declare_parameter("r_max", params_.r_max);
+  params_.max_fit_error =
+      this->declare_parameter("max_fit_error", params_.max_fit_error);
   // Params that need to be squared.
   double r_min, r_max, max_fit_error;
-  if (nh.getParam("r_min", r_min)) {
-    params.r_min_square = r_min*r_min;
+  if (this->get_parameter("r_min", r_min)) {
+    params_.r_min_square = r_min * r_min;
   }
-  if (nh.getParam("r_max", r_max)) {
-    params.r_max_square = r_max*r_max;
+  if (this->get_parameter("r_max", r_max)) {
+    params_.r_max_square = r_max * r_max;
   }
-  if (nh.getParam("max_fit_error", max_fit_error)) {
-    params.max_error_square = max_fit_error * max_fit_error;
+  if (this->get_parameter("max_fit_error", max_fit_error)) {
+    params_.max_error_square = max_fit_error * max_fit_error;
   }
+
+  segmenter_ = std::make_shared<GroundSegmentation>(params_);
 
   std::string ground_topic, obstacle_topic, input_topic;
-  bool latch;
-  nh.param<std::string>("input_topic", input_topic, "input_cloud");
-  nh.param<std::string>("ground_output_topic", ground_topic, "ground_cloud");
-  nh.param<std::string>("obstacle_output_topic", obstacle_topic, "obstacle_cloud");
-  nh.param("latch", latch, false);
+  ground_topic = this->declare_parameter("ground_output_topic", "ground_cloud");
+  obstacle_topic =
+      this->declare_parameter("obstacle_output_topic", "obstacle_cloud");
+  input_topic = this->declare_parameter("input_topic", "input_cloud");
+  cloud_sub_ = this->create_subscription<sensor_msgs::msg::PointCloud2>(
+      input_topic, rclcpp::SensorDataQoS(),
+      std::bind(&SegmentationNode::scanCallback, this, std::placeholders::_1));
+  ground_pub_ = this->create_publisher<sensor_msgs::msg::PointCloud2>(
+      ground_topic, 1000);
+  obstacle_pub_ = this->create_publisher<sensor_msgs::msg::PointCloud2>(
+      obstacle_topic, 1000);
+  tf_buffer_ = std::make_shared<tf2_ros::Buffer>(this->get_clock());
+  tf_listener_ = std::make_shared<tf2_ros::TransformListener>(*tf_buffer_);
+  RCLCPP_INFO(this->get_logger(), "Segmentation node initialized");
+  RCLCPP_INFO(this->get_logger(), "params_.r_min_square: %f",
+              params_.r_min_square);
+}
 
-  // Start node.
-  SegmentationNode node(nh, ground_topic, obstacle_topic, params, latch);
-  ros::Subscriber cloud_sub;
-  cloud_sub = nh.subscribe(input_topic, 1, &SegmentationNode::scanCallback, &node);
-  ros::spin();
+void SegmentationNode::scanCallback(
+    const sensor_msgs::msg::PointCloud2::SharedPtr msg) {
+  pcl::PointCloud<pcl::PointXYZ> cloud;
+  pcl::fromROSMsg(*msg, cloud);
+  pcl::PointCloud<pcl::PointXYZ> cloud_transformed;
+
+  std::vector<int> labels;
+
+  bool is_original_pc = true;
+  if (!gravity_aligned_frame_.empty()) {
+    geometry_msgs::msg::TransformStamped tf_stamped;
+    try {
+      tf_stamped = tf_buffer_->lookupTransform(
+          gravity_aligned_frame_, msg->header.frame_id, msg->header.stamp);
+      // Remove translation part.
+      tf_stamped.transform.translation.x = 0;
+      tf_stamped.transform.translation.y = 0;
+      tf_stamped.transform.translation.z = 0;
+      Eigen::Affine3d tf;
+      tf.translate(Eigen::Vector3d(0, 0, 0));
+      tf.rotate(Eigen::Quaterniond(
+          tf_stamped.transform.rotation.w, tf_stamped.transform.rotation.x,
+          tf_stamped.transform.rotation.y, tf_stamped.transform.rotation.z));
+      // tf::transformMsgToEigen(tf_stamped.transform, tf);
+      pcl::transformPointCloud(cloud, cloud_transformed, tf);
+      is_original_pc = false;
+    } catch (tf2::TransformException &ex) {
+      RCLCPP_WARN(this->get_logger(),
+                  "Failed to transform point cloud into "
+                  "gravity frame: %s",
+                  ex.what());
+    }
+  }
+
+  // Trick to avoid PC copy if we do not transform.
+  const pcl::PointCloud<pcl::PointXYZ> &cloud_proc =
+      is_original_pc ? cloud : cloud_transformed;
+
+  segmenter_->segment(cloud_proc, &labels);
+  pcl::PointCloud<pcl::PointXYZ> ground_cloud, obstacle_cloud;
+  for (size_t i = 0; i < cloud.size(); ++i) {
+    if (labels[i] == 1)
+      ground_cloud.push_back(cloud[i]);
+    else
+      obstacle_cloud.push_back(cloud[i]);
+  }
+  auto ground_msg = std::make_shared<sensor_msgs::msg::PointCloud2>();
+  auto obstacle_msg = std::make_shared<sensor_msgs::msg::PointCloud2>();
+  pcl::toROSMsg(ground_cloud, *ground_msg);
+  pcl::toROSMsg(obstacle_cloud, *obstacle_msg);
+  ground_msg->header = msg->header;
+  obstacle_msg->header = msg->header;
+  ground_pub_->publish(*ground_msg);
+  obstacle_pub_->publish(*obstacle_msg);
+}
+
+int main(int argc, char **argv) {
+  rclcpp::init(argc, argv);
+  rclcpp::NodeOptions options;
+  auto node = std::make_shared<SegmentationNode>(options);
+  rclcpp::spin(node);
+  rclcpp::shutdown();
+  return 0;
 }
